fixed bug that was rendering price twice in total
