+++ src/App.js
@@ -6,15 +6,6 @@ import AdditionalFeatures from "./components/AdditionalFeatures";
 import Total from "./components/Total";
 
 const App = () => {
-  // const state = {
-  //   additionalFeatures: [
-  //     { id: 1, name: "V-6 engine", price: 1500 },
-  //     { id: 2, name: "Racing detail package", price: 1500 },
-  //     { id: 3, name: "Premium sound system", price: 500 },
-  //     { id: 4, name: "Rear spoiler", price: 250 }
-  //   ]
-  // };
-
   return (
     <div className="boxes">
       <div className="box">
