+++ src/components/AdditionalFeature.js
@@ -3,15 +3,22 @@ import { connect } from "react-redux";
 import { addFeature } from "../actions";
 
 const AdditionalFeature = props => {
-  // const [total, setTotal] = useState(props.headerReducer.additionalPrice);
-  // console.log(total, "TOTAL AMOUNT");
-
   const onClick = e => {
     props.addFeature(Number(e.target.value));
+    // console.log(props.addFeature(Number(e.target.value)));
   };
+
+  // const checkChosen = () => {
+  //   let chosen = false;
+  //   props.additionalFeatures.forEach(item => {
+  //     if (item.id === props.feature.id) {
+  //       chosen = true;
+  //     }
+  //   });
+  //   return chosen;
+  // };
   return (
     <li>
-      {/* Add an onClick that will let you add a feature to your car */}
       <button value={props.feature.price} onClick={onClick} className="button">
         Add
       </button>
