+++ src/components/Total.js
@@ -4,20 +4,18 @@ import { connect } from "react-redux";
 export const Total = props => {
   return (
     <div className="content">
-      <h4>
-        Total Amount: $
-        {props.headerReducer.car.price + props.headerReducer.additionalPrice}
-      </h4>
+      <h4>Total Amount: ${props.car.price + props.additionalPrice}</h4>
     </div>
   );
 };
 
 const mapStateToProps = state => {
+  console.log(state);
   return {
     ...state,
     car: {
-      ...state.headerReducer.car,
-      price: state.headerReducer.car.price + state.headerReducer.additionalPrice
+      ...state.car,
+      price: state.car.price
     }
   };
 };
