+++ src/index.js
@@ -4,13 +4,13 @@ import App from "./App";
 
 import { createStore } from "redux";
 import { Provider } from "react-redux";
-import { rootReducer } from "./reducers";
+import { reducer } from "./reducers";
 
 import "bulma/css/bulma.css";
 import "./styles.scss";
 
 const store = createStore(
-  rootReducer,
+  reducer,
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
