+++ src/reducers/index.js
@@ -1,9 +1,38 @@
-import { combineReducers } from "redux";
+export const initialState = {
+  additionalPrice: 0,
+  car: {
+    price: 26395,
+    name: "2019 Ford Mustang",
+    image:
+      "https://cdn.motor1.com/images/mgl/0AN2V/s1/2019-ford-mustang-bullitt.jpg",
+    features: []
+  },
+  additionalFeatures: [
+    { id: 1, name: "V-6 engine", price: 1500 },
+    { id: 2, name: "Racing detail package", price: 1500 },
+    { id: 3, name: "Premium sound system", price: 500 },
+    { id: 4, name: "Rear spoiler", price: 250 }
+  ]
+};
+export const ADD_FEATURE = "ADD_FEATURE";
+// export const REMOVE_FEATURE = "REMOVE_FEATURE";
 
-import { headerReducer } from "./headerReducer";
-import { featuresReducer } from "./featuresReducer";
+export const reducer = (state = initialState, action) => {
+  // console.log(state, "reducer state");
+  switch (action.type) {
+    case ADD_FEATURE:
+      console.log(state, "REDUCER STATE");
+      return {
+        ...state,
+        additionalPrice: state.additionalPrice + action.payload
+      };
+    // case "REMOVE_FEATURE":
+    //   return {
+    //     ...state,
+    //     car: { ...state.car, feature: state.feature }
+    //   };
 
-export const rootReducer = combineReducers({
-  headerReducer,
-  featuresReducer
-});
+    default:
+      return state;
+  }
+};
